@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>DotVast.HashTool.WinUI.Core.Tests</RootNamespace>
 
     <IsPackable>false</IsPackable>
