@@ -7,14 +7,9 @@ using Microsoft.UI.Xaml;
 
 namespace DotVast.HashTool.WinUI.Activation;
 
-public sealed class DefaultActivationHandler : ActivationHandler<LaunchActivatedEventArgs>
+public sealed class DefaultActivationHandler(INavigationService navigationService) : ActivationHandler<LaunchActivatedEventArgs>
 {
-    private readonly INavigationService _navigationService;
-
-    public DefaultActivationHandler(INavigationService navigationService)
-    {
-        _navigationService = navigationService;
-    }
+    private readonly INavigationService _navigationService = navigationService;
 
     protected override bool CanHandleInternal(LaunchActivatedEventArgs args)
     {
