@@ -13,7 +13,7 @@ namespace DotVast.HashTool.WinUI.Enums;
 
 // Don't change the enum name unless absolutely necessary.
 // If the name is changed, it will destroy the deserialization and cause the app to start failure.
-[JsonConverter(typeof(JsonStringEnumConverter))]
+[JsonConverter(typeof(JsonStringEnumConverter<HashKind>))]
 public enum HashKind
 {
     CRC_32_V42,
@@ -95,8 +95,8 @@ internal static class HashKindExtensions
             HashKind.BLAKE2s_160 => HashLib4CSharp.Base.HashFactory.Crypto.CreateBlake2S_160().ToHashAlgorithm(),
             HashKind.BLAKE2s_224 => HashLib4CSharp.Base.HashFactory.Crypto.CreateBlake2S_224().ToHashAlgorithm(),
             HashKind.BLAKE2s_256 => HashLib4CSharp.Base.HashFactory.Crypto.CreateBlake2S_256().ToHashAlgorithm(),
-            HashKind.BLAKE2bp => HashLib4CSharp.Base.HashFactory.Crypto.CreateBlake2BP(64, Array.Empty<byte>()).ToHashAlgorithm(),
-            HashKind.BLAKE2sp => HashLib4CSharp.Base.HashFactory.Crypto.CreateBlake2SP(32, Array.Empty<byte>()).ToHashAlgorithm(),
+            HashKind.BLAKE2bp => HashLib4CSharp.Base.HashFactory.Crypto.CreateBlake2BP(64, []).ToHashAlgorithm(),
+            HashKind.BLAKE2sp => HashLib4CSharp.Base.HashFactory.Crypto.CreateBlake2SP(32, []).ToHashAlgorithm(),
             HashKind.BLAKE3 => new NativeCrypto.BLAKE3(),
             HashKind.XxHash32 => new System.IO.Hashing.XxHash32().ToHashAlgorithm(),
             HashKind.XxHash64 => new System.IO.Hashing.XxHash64().ToHashAlgorithm(),
