@@ -17,15 +17,15 @@ internal sealed partial class HashService : IHashService
 
     static readonly Dictionary<HashKind, HashMetadata> s_hashes = new()
     {
-        [HashKind.CRC_32_V42] = new("CRC-32", Aliases: new[] { "CRC-32/ADCCP", "CRC-32/V-42", "CRC-32/XZ", "PKZIP", "CRC-32/ISO-HDLC" }, IsEnabledForApp: true),
-        [HashKind.CRC_64_ECMA182] = new("CRC-64", Aliases: new[] { "CRC-64/ECMA-182" }, IsEnabledForApp: true),
+        [HashKind.CRC_32_V42] = new("CRC-32", Aliases: ["CRC-32/ADCCP", "CRC-32/V-42", "CRC-32/XZ", "PKZIP", "CRC-32/ISO-HDLC"], IsEnabledForApp: true),
+        [HashKind.CRC_64_ECMA182] = new("CRC-64", Aliases: ["CRC-64/ECMA-182"], IsEnabledForApp: true),
         [HashKind.MD4] = new("MD4"),
         [HashKind.MD5] = new("MD5", IsEnabledForApp: true, IsEnabledForContextMenu: true),
-        [HashKind.SHA1] = new("SHA-1", Aliases: new[] { "SHA1" }, IsEnabledForApp: true, IsEnabledForContextMenu: true),
-        [HashKind.SHA2_224] = new("SHA-224", Aliases: new[] { "SHA224" }),
-        [HashKind.SHA2_256] = new("SHA-256", Aliases: new[] { "SHA256" }, IsEnabledForApp: true, IsEnabledForContextMenu: true),
-        [HashKind.SHA2_384] = new("SHA-384", Aliases: new[] { "SHA384" }, IsEnabledForApp: true, IsEnabledForContextMenu: true),
-        [HashKind.SHA2_512] = new("SHA-512", Aliases: new[] { "SHA512" }, IsEnabledForApp: true, IsEnabledForContextMenu: true),
+        [HashKind.SHA1] = new("SHA-1", Aliases: ["SHA1"], IsEnabledForApp: true, IsEnabledForContextMenu: true),
+        [HashKind.SHA2_224] = new("SHA-224", Aliases: ["SHA224"]),
+        [HashKind.SHA2_256] = new("SHA-256", Aliases: ["SHA256"], IsEnabledForApp: true, IsEnabledForContextMenu: true),
+        [HashKind.SHA2_384] = new("SHA-384", Aliases: ["SHA384"], IsEnabledForApp: true, IsEnabledForContextMenu: true),
+        [HashKind.SHA2_512] = new("SHA-512", Aliases: ["SHA512"], IsEnabledForApp: true, IsEnabledForContextMenu: true),
         [HashKind.SHA3_224] = new("SHA3-224"),
         [HashKind.SHA3_256] = new("SHA3-256", IsEnabledForApp: true),
         [HashKind.SHA3_384] = new("SHA3-384", IsEnabledForApp: true),
@@ -42,22 +42,22 @@ internal sealed partial class HashService : IHashService
         [HashKind.BLAKE2bp] = new("BLAKE2bp"),
         [HashKind.BLAKE2sp] = new("BLAKE2sp", IsEnabledForApp: true),
         [HashKind.BLAKE3] = new("BLAKE3", IsEnabledForApp: true),
-        [HashKind.XxHash32] = new("XXH32", Aliases: new[] { "xxHash32" }, IsEnabledForApp: true, IsEnabledForContextMenu: true),
-        [HashKind.XxHash64] = new("XXH64", Aliases: new[] { "xxHash64" }, IsEnabledForApp: true, IsEnabledForContextMenu: true),
-        [HashKind.XxHash3] = new("XXH3", Aliases: new[] { "xxHash3" }, IsEnabledForApp: true),
-        [HashKind.XxHash128] = new("XXH128", Aliases: new[] { "xxHash128" }, IsEnabledForApp: true),
-        [HashKind.Keccak_224] = new("Keccak-224", Aliases: new[] { "Keccak224" }),
-        [HashKind.Keccak_256] = new("Keccak-256", Aliases: new[] { "Keccak256" }),
-        [HashKind.Keccak_288] = new("Keccak-288", Aliases: new[] { "Keccak288" }),
-        [HashKind.Keccak_384] = new("Keccak-384", Aliases: new[] { "Keccak384" }),
-        [HashKind.Keccak_512] = new("Keccak-512", Aliases: new[] { "Keccak512" }),
-        [HashKind.RIPEMD_128] = new("RIPEMD-128", Aliases: new[] { "RIPEMD128" }),
-        [HashKind.RIPEMD_160] = new("RIPEMD-160", Aliases: new[] { "RIPEMD160" }, IsEnabledForApp: true),
-        [HashKind.RIPEMD_256] = new("RIPEMD-256", Aliases: new[] { "RIPEMD256" }),
-        [HashKind.RIPEMD_320] = new("RIPEMD-320", Aliases: new[] { "RIPEMD320" }),
-        [HashKind.QuickXor] = new("QuickXor", Aliases: new[] { "QuickXorHash" }, Format: HashFormat.Base64),
+        [HashKind.XxHash32] = new("XXH32", Aliases: ["xxHash32"], IsEnabledForApp: true, IsEnabledForContextMenu: true),
+        [HashKind.XxHash64] = new("XXH64", Aliases: ["xxHash64"], IsEnabledForApp: true, IsEnabledForContextMenu: true),
+        [HashKind.XxHash3] = new("XXH3", Aliases: ["xxHash3"], IsEnabledForApp: true),
+        [HashKind.XxHash128] = new("XXH128", Aliases: ["xxHash128"], IsEnabledForApp: true),
+        [HashKind.Keccak_224] = new("Keccak-224", Aliases: ["Keccak224"]),
+        [HashKind.Keccak_256] = new("Keccak-256", Aliases: ["Keccak256"]),
+        [HashKind.Keccak_288] = new("Keccak-288", Aliases: ["Keccak288"]),
+        [HashKind.Keccak_384] = new("Keccak-384", Aliases: ["Keccak384"]),
+        [HashKind.Keccak_512] = new("Keccak-512", Aliases: ["Keccak512"]),
+        [HashKind.RIPEMD_128] = new("RIPEMD-128", Aliases: ["RIPEMD128"]),
+        [HashKind.RIPEMD_160] = new("RIPEMD-160", Aliases: ["RIPEMD160"], IsEnabledForApp: true),
+        [HashKind.RIPEMD_256] = new("RIPEMD-256", Aliases: ["RIPEMD256"]),
+        [HashKind.RIPEMD_320] = new("RIPEMD-320", Aliases: ["RIPEMD320"]),
+        [HashKind.QuickXor] = new("QuickXor", Aliases: ["QuickXorHash"], Format: HashFormat.Base64),
         [HashKind.Ed2k] = new("eD2k"),
-        [HashKind.HAS_160] = new("HAS-160", Aliases: new[] { "HAS160" }),
+        [HashKind.HAS_160] = new("HAS-160", Aliases: ["HAS160"]),
     };
 
     public string GetName(HashKind hashKind)
